Remove individual scan plans
