+++ src/crystallography_bluesky/i15_1/plans/__init__.py
@@ -4,15 +4,12 @@ from .robot import (
     robot_load,
     robot_unload,
 )
-from .take_eiger_data import take_eiger_and_i0_data, take_eiger_data
-from .take_i0_data import take_i0_data
+from .static_collection import static_collection_plan
 
 __all__ = [
     "robot_load",
     "move_hexapod_to_home_position",
     "prepare_beamline_for_robot_load",
     "robot_unload",
-    "take_eiger_data",
-    "take_i0_data",
-    "take_eiger_and_i0_data",
+    "static_collection_plan",
 ]
